+++ src/components/ResetPasswordForm.jsx
@@ -5,7 +5,7 @@ import ErrorMsg from "./ErrorMessage";
 import { GetResponseErrors } from "../utilities/CitysourcedResponseHelpers";
 import { Link } from 'react-router-dom';
 import FormContainer from './FormContainer';
-import { formIncomplete } from "../utilities/checkFormCompletion";
+import { IsFormComplete } from "../utilities/FormHelpers";
 import Alert from '../utilities/Alert';
 import SeButton from "./SeButton";
 import { ResetPassword } from './authService';
@@ -14,7 +14,7 @@ import { GoBack, Go, Routes, GoHome } from "../Routing";
 
 const ResetPasswordForm = (props, routeProps) => {
 
-	if (formIncomplete(props)) {
+	if (IsFormComplete(props)) {
 		GoHome(props);
 	}
