+++ src/components/SignInForm.jsx
@@ -6,7 +6,7 @@ import { GetResponseErrors } from "../utilities/CitysourcedResponseHelpers";
 import { Link } from 'react-router-dom';
 import FormContainer from './FormContainer';
 import { Login } from './authService';
-import { formIncomplete } from "../utilities/checkFormCompletion";
+import { IsFormComplete } from "../utilities/FormHelpers";
 import SeButton from "./SeButton";
 import { GoBack, GoHome, Go, Routes } from "../Routing";
 
@@ -20,7 +20,7 @@ const SignIn = (props, routeProps) => {
 		setFieldType(fieldType === 'Password' ? 'text' : 'Password');
 	};
 
-	if (formIncomplete(props)) {
+	if (IsFormComplete(props)) {
 		GoHome(props);
 	}
